moved validate_paths to io_utils
