+++ segmentation/utils/io_utils.py
@@ -1,5 +1,37 @@
 import os
+import pathlib
 import warnings
+
+
+def validate_paths(paths):
+    """Verifys that paths exist and don't leave Docker's scope
+
+    Args:
+        paths (str or list): paths to verify.
+
+    Raises:
+        ValueError: Raised if any directory is out of scope or non-existent
+    """
+
+    # if given a single path, convert to list
+    if not isinstance(paths, list):
+        paths = [paths]
+
+    for path in paths:
+        if not os.path.exists(path):
+            if str(path).startswith('../data'):
+                for parent in reversed(pathlib.Path(path).parents):
+                    if not os.path.exists(parent):
+                        raise ValueError(
+                            f'A bad path, {path}, was provided.\n'
+                            f'The folder, {parent.name}, could not be found...')
+                raise ValueError(
+                    f'The file/path, {pathlib.Path(path).name}, could not be found...')
+            else:
+                raise ValueError(
+                    f'The path, {path}, is not prefixed with \'../data\'.\n'
+                    f'Be sure to add all images/files/data to the \'data\' folder, '
+                    f'and to reference as \'../data/path_to_data/myfile.tif\'')
 
 
 def list_files(dir_name, substrs=None):
