+++ segmentation/utils/io_utils_test.py
@@ -1,8 +1,59 @@
 import os
 import tempfile
+import pathlib
 import pytest
 
 from segmentation.utils import io_utils as iou
+
+
+def test_validate_paths():
+
+    # change cwd to /scripts for more accurate testing
+    os.chdir('templates')
+
+    # make a tempdir for testing
+    with tempfile.TemporaryDirectory(dir='../data') as valid_path:
+
+        # make valid subdirectory
+        os.mkdir(valid_path + '/real_subdirectory')
+
+        # extract parts of valid path to alter for test cases
+        valid_parts = [p for p in pathlib.Path(valid_path).parts]
+        valid_parts[0] = 'not_a_real_directory'
+
+        # test no '../data' prefix
+        starts_out_of_scope = os.path.join(*valid_parts)
+
+        # construct test for bad middle folder path
+        valid_parts[0] = '..'
+        valid_parts[1] = 'data'
+        valid_parts[2] = 'not_a_real_subdirectory'
+        valid_parts.append('not_real_but_parent_is_problem')
+        bad_middle_path = os.path.join(*valid_parts)
+
+        # construct test for real path until file
+        wrong_file = os.path.join(valid_path + '/real_subdirectory', 'not_a_real_file.tiff')
+
+        # test one valid path
+        iou.validate_paths(valid_path)
+
+        # test multiple valid paths
+        iou.validate_paths([valid_path, '../data', valid_path + '/real_subdirectory'])
+
+        # test out-of-scope
+        with pytest.raises(ValueError, match=r".*not_a_real_directory.*prefixed.*"):
+            iou.validate_paths(starts_out_of_scope)
+
+        # test mid-directory existence
+        with pytest.raises(ValueError, match=r".*bad path.*not_a_real_subdirectory.*"):
+            iou.validate_paths(bad_middle_path)
+
+        # test file existence
+        with pytest.raises(ValueError, match=r".*The file/path.*not_a_real_file.*"):
+            iou.validate_paths(wrong_file)
+
+    # reset cwd after testing
+    os.chdir('../')
 
 
 def test_list_files():
